refactor: Menu, Restaurant 관련 코드 리팩토링
- user 관련 조회 기능 UserService로 코드 이동 후 사용하는 것으로 변경
- Restaurant, Menu에서 Campus 접근 로직을 추출하여 HasCampus 인터페이스로 통합(권한 검증 로직 통합)
