+++ src/main/java/sesac/server/group/entity/HasCampus.java
@@ -1,0 +1,8 @@
+package sesac.server.group.entity;
+
+import sesac.server.campus.entity.Campus;
+
+public interface HasCampus {
+
+    Campus getCampus();
+}
