+++ src/main/java/sesac/server/group/service/RestaurantService.java
@@ -20,15 +20,14 @@ import sesac.server.group.dto.response.RestaurantDetailResponse;
 import sesac.server.group.dto.response.RestaurantListForManagerResponse;
 import sesac.server.group.dto.response.RestaurantListResponse;
 import sesac.server.group.entity.GroupType;
+import sesac.server.group.entity.HasCampus;
 import sesac.server.group.entity.Menu;
 import sesac.server.group.entity.Restaurant;
 import sesac.server.group.exception.MenuErrorCode;
 import sesac.server.group.exception.RestaurantErrorCode;
 import sesac.server.group.repository.MenuRepository;
 import sesac.server.group.repository.RestaurantRepository;
-import sesac.server.user.exception.UserErrorCode;
-import sesac.server.user.repository.ManagerRepository;
-import sesac.server.user.repository.StudentRepository;
+import sesac.server.user.service.UserService;
 
 @Service
 @Log4j2
@@ -38,16 +37,14 @@ public class RestaurantService {
 
     private final RestaurantRepository restaurantRepository;
     private final MenuRepository menuRepository;
-    private final ManagerRepository managerRepository;
-    private final StudentRepository studentRepository;
+    private final UserService userService;
 
     // Restaurant
     public void createRestaurant(CustomPrincipal principal, GroupType groupType,
             CreateRestaurantRequest request) {
-        Campus campus = getManagerCampus(principal.id());
+        Campus campus = userService.getManagerCampus(principal.id());
 
-//        validateAddressInfo(request.addressId(), request.address(), request.longitude(),
-//                request.latitude());
+        // TODO: validateAddressInfo
 
         Restaurant restaurant = request.toEntity(campus, groupType);
 
@@ -57,7 +54,7 @@ public class RestaurantService {
     public List<RestaurantListForManagerResponse> getRestaurantListForManager(
             CustomPrincipal principal,
             GroupType type) {
-        Campus campus = getManagerCampus(principal.id());
+        Campus campus = userService.getManagerCampus(principal.id());
 
         return getRestaurantList(campus, type, RestaurantListForManagerResponse::from);
     }
@@ -65,7 +62,7 @@ public class RestaurantService {
     public List<RestaurantListResponse> getRestaurantList(
             CustomPrincipal principal,
             GroupType type) {
-        Campus campus = getStudentCampus(principal.id());
+        Campus campus = userService.getStudentCampus(principal.id());
 
         return getRestaurantList(campus, type, RestaurantListResponse::from);
     }
@@ -74,7 +71,7 @@ public class RestaurantService {
             GroupType type,
             Long restaurantId) {
         Restaurant restaurant = findRestaurantByIdAndType(restaurantId, type);
-        Campus userCampus = getUserCampus(principal);
+        Campus userCampus = userService.getUserCampus(principal);
         validateUserPermission(restaurant, userCampus);
 
         return RestaurantDetailResponse.from(restaurant);
@@ -83,9 +80,9 @@ public class RestaurantService {
     public void updateRestaurant(CustomPrincipal principal, GroupType type, Long restaurantId,
             UpdateRestaurantRequest request) {
         Restaurant restaurant = findRestaurantByIdAndType(restaurantId, type);
-        Campus managerCampus = getManagerCampus(principal.id());
-        validateUserPermission(restaurant, managerCampus);
+        Campus managerCampus = userService.getManagerCampus(principal.id());
 
+        validateUserPermission(restaurant, managerCampus);
         validateAddressUpdate(request);
 
         restaurant.update(request);
@@ -93,7 +90,7 @@ public class RestaurantService {
 
     public void deleteRestaurant(CustomPrincipal principal, GroupType type, Long restaurantId) {
         Restaurant restaurant = findRestaurantByIdAndType(restaurantId, type);
-        Campus managerCampus = getManagerCampus(principal.id());
+        Campus managerCampus = userService.getManagerCampus(principal.id());
         validateUserPermission(restaurant, managerCampus);
 
         restaurantRepository.delete(restaurant);
@@ -103,7 +100,7 @@ public class RestaurantService {
     public void createRestaurantMenu(CustomPrincipal principal, GroupType type,
             Long restaurantId, CreateMenuRequest request) {
         Restaurant restaurant = findRestaurantByIdAndType(restaurantId, type);
-        Campus managerCampus = getManagerCampus(principal.id());
+        Campus managerCampus = userService.getManagerCampus(principal.id());
         validateUserPermission(restaurant, managerCampus);
 
         // 메뉴 생성
@@ -116,14 +113,14 @@ public class RestaurantService {
         List<Menu> menu = menuRepository.findByRestaurantIdAndType(restaurantId, type);
         return menu.stream().map(MenuResponse::from).toList();
     }
-    
+
     public void updateRestaurantMenu(CustomPrincipal principal, GroupType type, Long restaurantId,
             Long menuId, UpdateMenuRequest request) {
         Menu menu = menuRepository.findByIdAndRestaurantIdAndType(menuId, restaurantId, type)
                 .orElseThrow(
                         () -> new BaseException(MenuErrorCode.NOT_FOUND_MENU)
                 );
-        Campus managerCampus = getManagerCampus(principal.id());
+        Campus managerCampus = userService.getManagerCampus(principal.id());
         validateUserPermission(menu, managerCampus);
 
         menu.update(request);
@@ -135,7 +132,7 @@ public class RestaurantService {
                 .orElseThrow(
                         () -> new BaseException(MenuErrorCode.NOT_FOUND_MENU)
                 );
-        Campus managerCampus = getManagerCampus(principal.id());
+        Campus managerCampus = userService.getManagerCampus(principal.id());
         validateUserPermission(menu, managerCampus);
 
         menuRepository.delete(menu);
@@ -155,14 +152,8 @@ public class RestaurantService {
                 .orElseThrow(() -> new BaseException(RestaurantErrorCode.NOT_FOUND_RESTAURANT));
     }
 
-    private void validateUserPermission(Restaurant restaurant, Campus userCampus) {
-        if (!restaurant.getCampus().equals(userCampus)) {
-            throw new BaseException(GlobalErrorCode.NO_PERMISSIONS);
-        }
-    }
-
-    private void validateUserPermission(Menu menu, Campus userCampus) {
-        if (!menu.getRestaurant().getCampus().equals(userCampus)) {
+    private void validateUserPermission(HasCampus entity, Campus userCampus) {
+        if (!entity.getCampus().equals(userCampus)) {
             throw new BaseException(GlobalErrorCode.NO_PERMISSIONS);
         }
     }
@@ -172,8 +163,7 @@ public class RestaurantService {
             if (!request.isAddressInfoComplete()) {
                 throw new BaseException(RestaurantErrorCode.INCOMPLETE_ADDRESS_INFO);
             }
-//            validateAddressInfo(request.addressId(), request.address(), request.longitude(),
-//                    request.latitude());
+            // TODO: validateAddressInfo
         }
     }
 
@@ -181,23 +171,5 @@ public class RestaurantService {
             BigDecimal latitude) {
         // TODO: 주소, 주소 id, 위도/경도 유효성 검사 구현
         return true;
-    }
-
-    private Campus getUserCampus(CustomPrincipal principal) {
-        return "STUDENT".equals(principal.role())
-                ? getStudentCampus(principal.id())
-                : getManagerCampus(principal.id());
-    }
-
-    private Campus getManagerCampus(Long managerId) {
-        return managerRepository.findById(managerId)
-                .orElseThrow(() -> new BaseException(UserErrorCode.NO_USER))
-                .getCampus();
-    }
-
-    private Campus getStudentCampus(Long studentId) {
-        return studentRepository.findById(studentId)
-                .orElseThrow(() -> new BaseException(UserErrorCode.NO_USER))
-                .getFirstCourse().getCampus();
     }
 }
